add configuration export option
