+++ musica/mechanism_configuration.cpp
@@ -374,7 +374,6 @@ void bind_mechanism_configuration(py::module_ & mechanism_configuration)
   py::class_<AqueousEquilibrium>(mechanism_configuration, "_AqueousEquilibrium")
       .def(py::init<>())
       .def_readwrite("name", &AqueousEquilibrium::name)
-      .def_readwrite("gas_phase", &AqueousEquilibrium::gas_phase)
       .def_readwrite("aerosol_phase", &AqueousEquilibrium::aerosol_phase)
       .def_readwrite("aerosol_phase_water", &AqueousEquilibrium::aerosol_phase_water)
       .def_readwrite("reactants", &AqueousEquilibrium::reactants)
