+++ musica/test/examples/v1/full_configuration/full_configuration.json
@@ -108,7 +108,8 @@
       "aerosol phase": "aqueous aerosol",
       "aerosol-phase species": "H2O2_aq",
       "aerosol-phase water": "H2O_aq",
-      "name": "my henry's law"
+      "name": "my henry's law",
+      "__irrelevant": "2"
     },
     {
       "type": "SIMPOL_PHASE_TRANSFER",
@@ -122,7 +123,8 @@
         1.96E-03,
         -4.96E-01
       ],
-      "name": "my simpol"
+      "name": "my simpol",
+      "__irrelevant": "2"
     },
     {
       "type": "AQUEOUS_EQUILIBRIUM",
@@ -147,7 +149,8 @@
           "coefficient": 1
         }
       ],
-      "name": "my aqueous eq"
+      "name": "my aqueous eq",
+      "__irrelevant": "2"
     },
     {
       "type": "CONDENSED_PHASE_ARRHENIUS",
@@ -174,7 +177,8 @@
           "coefficient": 1
         }
       ],
-      "name": "my condensed arrhenius"
+      "name": "my condensed arrhenius",
+      "__irrelevant": "2"
     },
     {
       "type": "CONDENSED_PHASE_ARRHENIUS",
@@ -220,7 +224,8 @@
         }
       ],
       "scaling factor": 12.3,
-      "name": "condensed photo B"
+      "name": "condensed photo B",
+      "__irrelevant": "2"
     },
     {
       "type": "EMISSION",
@@ -232,7 +237,8 @@
         }
       ],
       "name": "my emission",
-      "scaling factor": 12.3
+      "scaling factor": 12.3,
+      "__irrelevant": "2"
     },
     {
       "type": "FIRST_ORDER_LOSS",
@@ -244,7 +250,8 @@
         }
       ],
       "name": "my first order loss",
-      "scaling factor": 12.3
+      "scaling factor": 12.3,
+      "__irrelevant": "2"
     },
     {
       "type": "PHOTOLYSIS",
@@ -262,7 +269,8 @@
         }
       ],
       "name": "photo B",
-      "scaling factor": 12.3
+      "scaling factor": 12.3,
+      "__irrelevant": "2"
     },
     {
       "type": "SURFACE",
@@ -280,7 +288,8 @@
         }
       ],
       "aerosol phase": "surface reacting phase",
-      "name": "my surface"
+      "name": "my surface",
+      "__irrelevant": "2"
     },
     {
       "type": "TROE",
@@ -309,7 +318,8 @@
       "kinf_C": 24,
       "Fc": 0.9,
       "N": 0.8,
-      "name": "my troe"
+      "name": "my troe",
+      "__irrelevant": "2"
     },
     {
       "type": "BRANCHED_NO_RO2",
@@ -336,7 +346,8 @@
       "Y": 167,
       "a0": 0.15,
       "n": 9,
-      "name": "my branched"
+      "name": "my branched",
+      "__irrelevant": "2"
     },
     {
       "gas phase": "gas",
@@ -356,13 +367,15 @@
           "species name": "C",
           "coefficient": 1
         }
-      ]
+      ],
+      "__irrelevant": "2"
     },
     {
       "type": "WET_DEPOSITION",
       "aerosol phase": "cloud",
       "name": "rxn cloud",
-      "scaling factor": 12.3
+      "scaling factor": 12.3,
+      "__irrelevant": "2"
     },
     {
       "type": "ARRHENIUS",
@@ -384,7 +397,8 @@
       "C": 102.3,
       "D": 63.4,
       "E": -1.3,
-      "name": "my arrhenius"
+      "name": "my arrhenius",
+      "__irrelevant": "2"
     },
     {
       "type": "ARRHENIUS",
@@ -428,7 +442,8 @@
         }
       ],
       "name": "my user defined",
-      "scaling factor": 12.3
+      "scaling factor": 12.3,
+      "__irrelevant": "2"
     }
   ]
 }
