+++ pyproject.toml
@@ -18,6 +18,7 @@ maintainers = [{ name = "ACOM MUSICA Developers", email = "musica-support@ucar.e
 license = { file = "LICENSE" }
 
 dependencies = [
+    "pyyaml>=6.0.2",
 ]
 
 [project.optional-dependencies]
